Update build definition for jmx-memory
This adopts the settings that are coming for other plugins as well as
a result of the main build system overhaul, in which renaissance-core
does not have a private build definition.

We specifically target JDK8, get rid of the sbt-assembly plugin
(because it is not needed), list renaissance-core as provided
dependency, and set the organization name manually.